Extract executeJob into dedicated module to break circular dependency
Move executeJob, its helpers (parseContinuationTag, loadPlanNote), and
related constants (system prompts, MAX_RETRIES, RETRY_DELAY_MS) from
heartbeat.ts into src/cron/execute-job.ts.

This breaks the circular dependency cycle:
  jobs.ts → heartbeat.ts → slack.ts → jobs.ts

Now heartbeat.ts and jobs.ts both import executeJob from execute-job.ts,
and heartbeat.ts no longer participates in any import cycle.
